Update script gql endpoints
- Update the GraphQL endpoints for the scripts
- Move `MerkleDropClaims` and `ScriptRewards` queries so that they are run on the legacy subgraph
- Update script types
- Remove now-redundant fees breakdown script
