+++ scripts/utils/getApolloClient.ts
@@ -35,7 +35,9 @@ export const getApolloClient = () => {
   const apolloLink = ApolloLink.from([
     new MultiAPILink({
       endpoints: {
-        mstable: process.env.REACT_APP_GRAPHQL_ENDPOINT_PROTOCOL as string,
+        legacy: process.env.REACT_APP_GRAPHQL_ENDPOINT_LEGACY as string,
+        protocol: process.env.REACT_APP_GRAPHQL_ENDPOINT_PROTOCOL as string,
+        ecosystem: process.env.REACT_APP_GRAPHQL_ENDPOINT_ECOSYSTEM as string,
         balancer: process.env.REACT_APP_GRAPHQL_ENDPOINT_BALANCER as string,
         uniswap: process.env.REACT_APP_GRAPHQL_ENDPOINT_UNISWAP as string,
         blocks: process.env.REACT_APP_GRAPHQL_ENDPOINT_BLOCKS as string,
