+++ src/graphql/ecosystem.tsx
@@ -1857,34 +1857,6 @@ export type RewardsDistributorQueryVariables = {};
 
 export type RewardsDistributorQuery = { rewardsDistributors: Array<Pick<RewardsDistributor, 'id' | 'fundManagers'>> };
 
-export type MerkleDropClaimsQueryVariables = {
-  account: Scalars['Bytes'];
-};
-
-
-export type MerkleDropClaimsQuery = { merkleDrops: Array<(
-    Pick<MerkleDrop, 'id'>
-    & { token: TokenDetailsFragment, tranches: Array<(
-      Pick<MerkleDropTranche, 'trancheNumber' | 'totalAmount'>
-      & { claims: Array<Pick<MerkleDropClaim, 'balance'>> }
-    )> }
-  )> };
-
-export type ScriptRewardsQueryVariables = {
-  id: Scalars['ID'];
-  end: Scalars['BigInt'];
-  block?: Maybe<Block_Height>;
-  limit: Scalars['Int'];
-  offset: Scalars['Int'];
-};
-
-
-export type ScriptRewardsQuery = { stakingRewardsContracts: Array<(
-    Pick<StakingRewardsContract, 'lastUpdateTime' | 'periodFinish' | 'rewardPerTokenStored' | 'rewardRate' | 'totalSupply'>
-    & { address: StakingRewardsContract['id'] }
-    & { stakingRewards: Array<Pick<StakingReward, 'amount' | 'account' | 'amountPerTokenPaid'>>, stakingBalances: Array<Pick<StakingBalance, 'amount' | 'account'>>, claimRewardTransactions: Array<Pick<StakingRewardsContractClaimRewardTransaction, 'amount' | 'sender'>> }
-  )> };
-
 export const TokenDetailsFragmentDoc = gql`
     fragment TokenDetails on Token {
   id
@@ -2088,101 +2060,3 @@ export function useRewardsDistributorLazyQuery(baseOptions?: ApolloReactHooks.La
 export type RewardsDistributorQueryHookResult = ReturnType<typeof useRewardsDistributorQuery>;
 export type RewardsDistributorLazyQueryHookResult = ReturnType<typeof useRewardsDistributorLazyQuery>;
 export type RewardsDistributorQueryResult = ApolloReactCommon.QueryResult<RewardsDistributorQuery, RewardsDistributorQueryVariables>;
-export const MerkleDropClaimsDocument = gql`
-    query MerkleDropClaims($account: Bytes!) @api(name: ecosystem) {
-  merkleDrops {
-    id
-    token {
-      ...TokenDetails
-    }
-    tranches(orderDirection: asc, orderBy: trancheNumber, where: {expired: false}) {
-      trancheNumber
-      totalAmount
-      claims(where: {account: $account}) {
-        balance
-      }
-    }
-  }
-}
-    ${TokenDetailsFragmentDoc}`;
-
-/**
- * __useMerkleDropClaimsQuery__
- *
- * To run a query within a React component, call `useMerkleDropClaimsQuery` and pass it any options that fit your needs.
- * When your component renders, `useMerkleDropClaimsQuery` returns an object from Apollo Client that contains loading, error, and data properties
- * you can use to render your UI.
- *
- * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
- *
- * @example
- * const { data, loading, error } = useMerkleDropClaimsQuery({
- *   variables: {
- *      account: // value for 'account'
- *   },
- * });
- */
-export function useMerkleDropClaimsQuery(baseOptions?: ApolloReactHooks.QueryHookOptions<MerkleDropClaimsQuery, MerkleDropClaimsQueryVariables>) {
-        return ApolloReactHooks.useQuery<MerkleDropClaimsQuery, MerkleDropClaimsQueryVariables>(MerkleDropClaimsDocument, baseOptions);
-      }
-export function useMerkleDropClaimsLazyQuery(baseOptions?: ApolloReactHooks.LazyQueryHookOptions<MerkleDropClaimsQuery, MerkleDropClaimsQueryVariables>) {
-          return ApolloReactHooks.useLazyQuery<MerkleDropClaimsQuery, MerkleDropClaimsQueryVariables>(MerkleDropClaimsDocument, baseOptions);
-        }
-export type MerkleDropClaimsQueryHookResult = ReturnType<typeof useMerkleDropClaimsQuery>;
-export type MerkleDropClaimsLazyQueryHookResult = ReturnType<typeof useMerkleDropClaimsLazyQuery>;
-export type MerkleDropClaimsQueryResult = ApolloReactCommon.QueryResult<MerkleDropClaimsQuery, MerkleDropClaimsQueryVariables>;
-export const ScriptRewardsDocument = gql`
-    query ScriptRewards($id: ID!, $end: BigInt!, $block: Block_height, $limit: Int!, $offset: Int!) @api(name: ecosystem) {
-  stakingRewardsContracts(where: {id: $id}, block: $block) {
-    address: id
-    lastUpdateTime
-    periodFinish
-    rewardPerTokenStored
-    rewardRate
-    totalSupply
-    stakingRewards(where: {type: REWARD}, first: $limit, skip: $offset) {
-      amount
-      account
-      amountPerTokenPaid
-    }
-    stakingBalances(first: $limit, skip: $offset) {
-      amount
-      account
-    }
-    claimRewardTransactions(first: $limit, skip: $offset, orderBy: timestamp, orderDirection: asc, where: {timestamp_lt: $end}) {
-      amount
-      sender
-    }
-  }
-}
-    `;
-
-/**
- * __useScriptRewardsQuery__
- *
- * To run a query within a React component, call `useScriptRewardsQuery` and pass it any options that fit your needs.
- * When your component renders, `useScriptRewardsQuery` returns an object from Apollo Client that contains loading, error, and data properties
- * you can use to render your UI.
- *
- * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
- *
- * @example
- * const { data, loading, error } = useScriptRewardsQuery({
- *   variables: {
- *      id: // value for 'id'
- *      end: // value for 'end'
- *      block: // value for 'block'
- *      limit: // value for 'limit'
- *      offset: // value for 'offset'
- *   },
- * });
- */
-export function useScriptRewardsQuery(baseOptions?: ApolloReactHooks.QueryHookOptions<ScriptRewardsQuery, ScriptRewardsQueryVariables>) {
-        return ApolloReactHooks.useQuery<ScriptRewardsQuery, ScriptRewardsQueryVariables>(ScriptRewardsDocument, baseOptions);
-      }
-export function useScriptRewardsLazyQuery(baseOptions?: ApolloReactHooks.LazyQueryHookOptions<ScriptRewardsQuery, ScriptRewardsQueryVariables>) {
-          return ApolloReactHooks.useLazyQuery<ScriptRewardsQuery, ScriptRewardsQueryVariables>(ScriptRewardsDocument, baseOptions);
-        }
-export type ScriptRewardsQueryHookResult = ReturnType<typeof useScriptRewardsQuery>;
-export type ScriptRewardsLazyQueryHookResult = ReturnType<typeof useScriptRewardsLazyQuery>;
-export type ScriptRewardsQueryResult = ApolloReactCommon.QueryResult<ScriptRewardsQuery, ScriptRewardsQueryVariables>;
