+++ src/graphql/ecosystem/query.graphql
@@ -86,62 +86,3 @@ query RewardsDistributor @api(name: ecosystem) {
         fundManagers
     }
 }
-
-query MerkleDropClaims($account: Bytes!) @api(name: ecosystem) {
-    merkleDrops {
-        id
-        token {
-            ...TokenDetails
-        }
-        tranches(
-            orderDirection: asc
-            orderBy: trancheNumber
-            where: { expired: false }
-        ) {
-            trancheNumber
-            totalAmount
-            claims(where: { account: $account }) {
-                balance
-            }
-        }
-    }
-}
-
-query ScriptRewards(
-    $id: ID!
-    $end: BigInt!
-    $block: Block_height
-    $limit: Int!
-    $offset: Int!
-) @api(name: ecosystem) {
-    stakingRewardsContracts(where: { id: $id }, block: $block) {
-        address: id
-        lastUpdateTime
-        periodFinish
-        rewardPerTokenStored
-        rewardRate
-        totalSupply
-
-        stakingRewards(where: { type: REWARD }, first: $limit, skip: $offset) {
-            amount
-            account
-            amountPerTokenPaid
-        }
-
-        stakingBalances(first: $limit, skip: $offset) {
-            amount
-            account
-        }
-
-        claimRewardTransactions(
-            first: $limit
-            skip: $offset
-            orderBy: timestamp
-            orderDirection: asc
-            where: { timestamp_lt: $end }
-        ) {
-            amount
-            sender
-        }
-    }
-}
